requirements.txt
sounddevice
scipy
openai-whisper
transformers
pandas
pygame
flask-cors